requirements.txt
backcall==0.2.0
beautifulsoup4==4.10.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
coverage==6.1.2
cycler==0.11.0
debugpy==1.5.1
decorator==5.1.1
entrypoints==0.4
filelock==3.5.0
fonttools==4.29.1
funcy==1.17
future==0.18.2
gensim==4.1.2
idna==3.3
ipykernel==6.6.0
ipython==7.30.1
jedi==0.18.1
Jinja2==3.0.3
joblib==1.1.0
jupyter-client==7.1.0
jupyter-core==4.9.1
kiwisolver==1.3.2
lxml==4.7.1
MarkupSafe==2.0.1
matplotlib==3.5.0
matplotlib-inline==0.1.3
nest-asyncio==1.5.4
nltk==3.6.5
numexpr==2.8.1
numpy==1.21.4
packaging==21.3
pandas==1.3.5
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2-binary==2.9.2
-e .
ptyprocess==0.7.0
Pygments==2.11.2
pyLDAvis==3.3.1
pyparsing==3.0.7
python-dateutil==2.8.2
pytz==2021.3
PyYAML<6.0
pyzmq==22.3.0
regex==2022.1.18
requests==2.26.0
scikit-learn==1.0.2
scipy==1.7.3
setuptools==60.9.1
setuptools-scm==6.4.2
six==1.16.0
sklearn==0.0
smart-open==5.2.1
snscrape==0.4.3.20220106
soupsieve==2.3.1
threadpoolctl==3.1.0
tomli==2.0.1
tomotopy==0.12.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
urllib3==1.26.8
wcwidth==0.2.5
yapf==0.31.0
pytrends==4.8.0
